Port code to use rustix instead of ioctl-sys
This commit replaces the code using ioctl-sys with equivalent code using
rustix. In addition to safety, this move also removes the dependency on
C code and replaces it with a syscall, which is much faster.

This only applies for Linux, as macOS's "clonefile" call has not been ported
to rustix yet.

Signed-off-by: John Nunley <dev@notgull.net>